+++ disasm-test/Main.hs
@@ -398,6 +398,7 @@ assemblyCube rootPath llvmver = TS.mkCUBE
                                            , "pre-llvm16"
                                            , "pre-llvm17"
                                            , "post-llvm18"
+                                           , "pre-llvm19"
                                            ])
                      ]
     -- Somewhat unusually for tasty-sugar, we make the expectedSuffix the same